prototype finished
for now this is where I'll leave this project, maybe redo it in Qwik with typescript at some point.

The background changes as well as the icon based on the weathercode, so does all the data.

The custom option select is a bit buggy, but for now that will do.